requirements.txt
anyio==3.2.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
orjson==3.5.3
Pillow==8.2.0
pycryptodome==3.10.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
urllib3==1.26.5
